Change the for-loop during encryption to a do-while.
We want to be able to encrypt zero-length plaintexts, but the current
implementation would also append a tag-only record (17 octets) when
the record *exactly* fits.

(For example, rs=1024, padding=0, data_size=1023).

This would throw an error about a truncated payload during the decryption
phase, which I removed as well. The drafts don't clarify what the
truncation behavior of records should be.